Use a one second timeout when checking for version
Checking versions is a "nice to have" sort of thing, so we don't want to
block progress elsewhere when the checkpoint API is unreachable for any
reason.

This leads to an error being returned if the timeout is hit. Our callers
that use this API to check for new versions just write errors to the log
and continue, so this causes the version check to effectively be ignored
when the checkpoint API is down.